First version of dot autosubst output
This gives dummy declarations of gname and label, which I will have to fix by
hand later.

In fact, the correct definition of `vobj` is `vobj : (vl -> dms) -> vl`, not
`vobj: dms -> vl`, but this is rejected (IMHO incorrectly) by the Autosubst2
prototype, with error

```
The type vl contains a vacuous binding of a variable of type vl in constructor
vobj.
```

The reason appears to be that Autosubst2 demands a direct occurrence of `vl` in
`dms`, and is not satisfied  by the indirect occurrence going through
`dcons (dvl ...)`.