Fix Autosubst2 to handle Dot
The fix is taken from 96184e64e31f54ae0bc0b4da786ce337b0a23cc6. This time, this
fix is sufficient alone to produce compiling code.

Without this fix, Autosubst2 gives these spurious errors:

```
The type ty contains a vacuous binding of a variable of type vl in constructor TAll.
The type ty contains a vacuous binding of a variable of type vl in constructor TMu.
The type vl contains a vacuous binding of a variable of type vl in constructor vobj.
```